fix: third wave of clang -Werror sign/include/unused fixes
Iterating against an actual Ubuntu 24.04 + clang 18 build:

- engine/codepage.cpp utf8_to_cp437: cast every `byte & mask`
  intermediate to uint32_t explicitly to silence
  -Wsign-conversion on the implicit int → uint32_t step.
- drivers/cdx/cdx_driver apply_ctr_: bump `b` from int to
  std::size_t inside the counter-increment loop.
- drivers/ntx/ntx_index next/prev cache lookups + format_empty_page
  loop + insert(begin+pos) splice + ents[pos+1] subscript +
  pmid splice: cast every signed index to size_t / ptrdiff_t.
- drivers/cdx/cdx_index next/prev cur_index_ subscripts: bulk
  sed cur_decoded_[cur_index_] → cur_decoded_[size_t-cast].
- tests/unit/abi_encryption_test, network_server_test: switch
  the inner `for (int i = 0; i < 4/5; ++i)` loops over tag
  bytes to size_t.
- tests/unit/abi_aep_test: missing #include <vector> on Linux
  (Win MSVC ships it transitively via filesystem).
- tests/unit/tx_log_test: missing #include <algorithm> for
  std::sort + size_t loop counter.
- tests/unit/dbf_vfp_test write_u16_le + abi/ace_exports
  find_field: tag with [[maybe_unused]] (only used on certain
  paths; clang flags the unused TU symbol under -Wunused-function).

Win build still green: 334 / 5036.

+++ src/drivers/cdx/cdx_index.cpp
@@ -476,8 +476,8 @@ util::Result<SeekOutcome> CdxIndex::seek_last() {
         return SeekOutcome{SeekHit::AfterEnd, 0, false};
     }
     cur_index_ = static_cast<std::int32_t>(cur_decoded_.size() - 1);
-    current_key_ = cur_decoded_[cur_index_].first;
-    return SeekOutcome{SeekHit::Exact, cur_decoded_[cur_index_].second, true};
+    current_key_ = cur_decoded_[static_cast<std::size_t>(cur_index_)].first;
+    return SeekOutcome{SeekHit::Exact, cur_decoded_[static_cast<std::size_t>(cur_index_)].second, true};
 }
 
 util::Result<SeekOutcome>
@@ -526,8 +526,8 @@ util::Result<SeekOutcome> CdxIndex::next() {
     if (cur_index_ < 0) return SeekOutcome{SeekHit::AfterEnd, 0, false};
     if (static_cast<std::size_t>(cur_index_ + 1) < cur_decoded_.size()) {
         cur_index_ += 1;
-        current_key_ = cur_decoded_[cur_index_].first;
-        return SeekOutcome{SeekHit::Exact, cur_decoded_[cur_index_].second, true};
+        current_key_ = cur_decoded_[static_cast<std::size_t>(cur_index_)].first;
+        return SeekOutcome{SeekHit::Exact, cur_decoded_[static_cast<std::size_t>(cur_index_)].second, true};
     }
     auto pg = get_page_(cur_leaf_);
     if (!pg) return pg.error();
@@ -553,8 +553,8 @@ util::Result<SeekOutcome> CdxIndex::prev() {
     if (cur_index_ < 0) return SeekOutcome{SeekHit::BeforeBegin, 0, false};
     if (cur_index_ > 0) {
         cur_index_ -= 1;
-        current_key_ = cur_decoded_[cur_index_].first;
-        return SeekOutcome{SeekHit::Exact, cur_decoded_[cur_index_].second, true};
+        current_key_ = cur_decoded_[static_cast<std::size_t>(cur_index_)].first;
+        return SeekOutcome{SeekHit::Exact, cur_decoded_[static_cast<std::size_t>(cur_index_)].second, true};
     }
     auto pg = get_page_(cur_leaf_);
     if (!pg) return pg.error();
@@ -572,8 +572,8 @@ util::Result<SeekOutcome> CdxIndex::prev() {
         return SeekOutcome{SeekHit::BeforeBegin, 0, false};
     }
     cur_index_ = static_cast<std::int32_t>(cur_decoded_.size() - 1);
-    current_key_ = cur_decoded_[cur_index_].first;
-    return SeekOutcome{SeekHit::Exact, cur_decoded_[cur_index_].second, true};
+    current_key_ = cur_decoded_[static_cast<std::size_t>(cur_index_)].first;
+    return SeekOutcome{SeekHit::Exact, cur_decoded_[static_cast<std::size_t>(cur_index_)].second, true};
 }
 
 util::Result<void>
